# CUDA: skip masked out KQ slices in mma FA kernel