This PR extends the mma-based CUDA FlashAttention kernel with logic for skipping fully masked-out KQ slices. However, this kernel makes use of asynchronous data loading to preload KV data so this is not straightforward: the mask and K data are being preloaded at the same time, if it turns out that the KQ slice should be skipped the K I/O is wasted, and also the GPU compute pipes are idling until the mask and K data for the next potential KQ slice can be fetched. It turns out that for the new batched inference setup it's faster not to preload any data at all but this is going to hamper the overall performance. With LLaMA 3 8b q4_0 and the command

```
./batched-bench -t 8 -tb 8 -m models/opt/${model_name}-${quantization}.gguf -c 65536 -b 1024 -ub 1024 -npp 1024 -ntg 128 -npl 1,2,4,8,16,32 -fa -ngl 999
```

I'm seeing net speedups of ~3% for my RTX 3090/4090, with

```
LLAMA_ARG_MODEL=/opt/models/llama_3.2_instruct-1b-q4_k_m.gguf py server-bench.py --path_server /home/johannesg/Projects/llama.cpp/build/bin/llama-server --prompt_source rng-1024-40000
```

I'm seeing net speedups of ~10% (`LLAMA_SET_ROWS=1` set for both cases).

I think that the approach of loading a tile of the mask and checking whether it's all `== -inf` is bad. I think I could write a much better kernel if I instead had a list of those 256x64 mask slices that are not all `== -inf`. Then I could simply iterate over those indices and preload data without potentially wasting any I/O. It would also allow me to solve the problem where I cannot distribute work to streaming multiprocessors in an optimal way because I don't know ahead of time how the KQ slices needing compute are distributed; if one of the sequences is very long vs. the rest I would ideally assign more SMs to that sequence. For non-batched inference with a diagonal mask I could also skip a few KQ slices, it would reduce the amount of KV data to iterate over per token by half the physical batch size (probably not worthwhile on its own). Ideally one bit of the indices should be reserved to indicate whether all elements in the mask slice are `== 0`, the mask for those slices will not need to be loaded (as of yet unclear whether that would make a meaningful difference). @ggerganov thoughts?