@JohannesGaessler 2025-07-17
>What we can do is during the construction of the KQ mask, we can create a second tensor that contains non-masked block indices.

Yes, that is how I would have implemented it as well.

>We could also pass a bitmask if that would be more appropriate.

For CUDA a bitmask would not work well because I don't know where in the bit mask I would have to set the start and end indices for evenly distributing data across streaming multiprocessors.

>The first test that you did with llama-batched-bench does not benefit at all from the -INF masking optimization because the sequences are always of equal length. But it is performing faster without preloading the mask - that's unexpected. Could this pre-loading not have a significant effect?

For a single sequence the preloading definitely does provide a benefit, the difference is something like 10% end-to-end speedup. Also I just noticed that I forgot that master and the PR are running different kernels for the RTX 4090, master is running the vector kernel, the PR is running the mma kernel. There is also a difference for the RTX 3090, what could be happening there is that due to the lower shared memory use for a single stage pipeline there is some benefit from higher occupancy.

>Can you post the resulting numbers from running the llama-batched-bench on master and this PR to see how the respective numbers change?

<details>

master, RTX 4090:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|  1024 |    128 |    1 |   1152 |    0.083 | 12306.07 |    0.823 |   155.61 |    0.906 |  1271.84 |
|  1024 |    128 |    2 |   2304 |    0.153 | 13425.63 |    0.922 |   277.69 |    1.074 |  2144.39 |
|  1024 |    128 |    4 |   4608 |    0.307 | 13352.46 |    0.990 |   517.06 |    1.297 |  3552.90 |
|  1024 |    128 |    8 |   9216 |    0.613 | 13371.35 |    1.274 |   803.66 |    1.887 |  4884.38 |
|  1024 |    128 |   16 |  18432 |    1.228 | 13344.11 |    1.668 |  1227.51 |    2.896 |  6364.14 |
|  1024 |    128 |   32 |  36864 |    2.463 | 13306.32 |    2.304 |  1777.46 |    4.767 |  7733.16 |

PR, RTX 4090:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|  1024 |    128 |    1 |   1152 |    0.084 | 12165.57 |    0.823 |   155.51 |    0.907 |  1269.75 |
|  1024 |    128 |    2 |   2304 |    0.153 | 13403.58 |    0.936 |   273.58 |    1.089 |  2116.59 |
|  1024 |    128 |    4 |   4608 |    0.307 | 13328.60 |    1.002 |   511.02 |    1.309 |  3519.66 |
|  1024 |    128 |    8 |   9216 |    0.614 | 13333.18 |    1.282 |   798.63 |    1.897 |  4859.21 |
|  1024 |    128 |   16 |  18432 |    1.231 | 13313.54 |    1.675 |  1222.85 |    2.905 |  6344.04 |
|  1024 |    128 |   32 |  36864 |    2.469 | 13272.45 |    2.286 |  1791.48 |    4.755 |  7752.27 |

master, RTX 3090:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|  1024 |    128 |    1 |   1152 |    0.192 |  5336.50 |    0.990 |   129.33 |    1.182 |   974.96 |
|  1024 |    128 |    2 |   2304 |    0.374 |  5482.68 |    1.108 |   231.05 |    1.482 |  1555.14 |
|  1024 |    128 |    4 |   4608 |    0.751 |  5450.97 |    1.345 |   380.63 |    2.097 |  2197.87 |
|  1024 |    128 |    8 |   9216 |    1.522 |  5380.98 |    2.224 |   460.37 |    3.747 |  2459.77 |
|  1024 |    128 |   16 |  18432 |    3.058 |  5358.17 |    2.308 |   887.22 |    5.366 |  3434.89 |
|  1024 |    128 |   32 |  36864 |    6.143 |  5334.01 |    3.322 |  1232.96 |    9.465 |  3894.64 |

PR, RTX 3090:

|    PP |     TG |    B |   N_KV |   T_PP s | S_PP t/s |   T_TG s | S_TG t/s |      T s |    S t/s |
|-------|--------|------|--------|----------|----------|----------|----------|----------|----------|
|  1024 |    128 |    1 |   1152 |    0.194 |  5270.15 |    0.993 |   128.89 |    1.187 |   970.17 |
|  1024 |    128 |    2 |   2304 |    0.381 |  5373.56 |    1.133 |   225.92 |    1.514 |  1521.55 |
|  1024 |    128 |    4 |   4608 |    0.761 |  5382.02 |    1.350 |   379.32 |    2.111 |  2183.00 |
|  1024 |    128 |    8 |   9216 |    1.534 |  5339.30 |    2.240 |   457.08 |    3.775 |  2441.60 |
|  1024 |    128 |   16 |  18432 |    3.089 |  5304.12 |    2.310 |   886.48 |    5.399 |  3413.85 |
|  1024 |    128 |   32 |  36864 |    6.201 |  5284.08 |    3.209 |  1276.23 |    9.411 |  3917.24 |

</details>

I'll make a prototype for a kernel using a list of mask slices.