@JohannesGaessler 2025-08-13
Obsoleted by https://github.com/ggml-org/llama.cpp/pull/14924